requirements.txt
streamlit==1.36.0
langgraph>=0.2.40
langchain-core>=0.3.0
langchain-groq>=0.1.0
pydantic>=2.7
python-dotenv>=1.0
requests>=2.32
tavily-python>=0.3.4
tzdata>=2024.1       